Fixed User addition and deletion

+++ social-network/src/main/java/spring/summer/socialnetwork/services/UserService.java
@@ -2,6 +2,8 @@ package spring.summer.socialnetwork.services;
 
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import spring.summer.socialnetwork.dto.UserDTO;
@@ -9,6 +11,7 @@ import spring.summer.socialnetwork.models.User;
 import spring.summer.socialnetwork.repositories.UserRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -22,24 +25,36 @@ public class UserService {
     }
 
     @Transactional
-    public void saveUser(UserDTO userDTO) {
-        var user = User.builder()
-                .name(userDTO.getName())
-                .surname(userDTO.getSurname())
-                .email(userDTO.getEmail())
-                .password(passwordEncoder.encode(userDTO.getPassword()))
-                .build();
+    public ResponseEntity<User> saveUser(UserDTO userDTO) {
+        var user = User.builder().name(userDTO.getName()).surname(userDTO.getSurname()).email(userDTO.getEmail()).password(passwordEncoder.encode(userDTO.getPassword())).build();
 
         userRepository.save(user);
+
+        return ResponseEntity.ok(user);
+
     }
 
     @Transactional
-    public void deleteUser(String id) {
-        userRepository.deleteById(Long.parseLong(id));
+    public ResponseEntity deleteUser(String id) {
+        if (userRepository.findById(Long.parseLong(id)).equals(Optional.empty()))
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        else {
+            userRepository.deleteById(Long.parseLong(id));
+            return ResponseEntity.status(HttpStatus.ACCEPTED).build();
+        }
     }
 
     public List<User> getAllUsers() {
         return userRepository.findAll();
+    }
+
+    public ResponseEntity<User> getUserById(String id) {
+        if (userRepository.findById(Long.parseLong(id)).equals(Optional.empty()))
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        else {
+            var user = userRepository.findById(Long.parseLong(id)).get();
+            return ResponseEntity.status(HttpStatus.ACCEPTED).body(user);
+        }
     }
 
     @Transactional
